fix(ee): burn the gosec backlog from 145 to 15, no waivers
Finishes the ee/ corpus. 131 actionable findings fixed across 33 packages, with
ZERO #nosec and ZERO //nolint added anywhere (verified by diffing for both), and
zero t.Skip introduced.

Fanned out one agent per package, each behind an independent verifier that
re-ran the lint, the build and the tests itself rather than trusting the fixing
agent's report, and read the diff specifically to judge whether any test had
been weakened or any conversion's observable behaviour changed. 37 fix agents
returned build_ok and tests_ok with waivers_added=0.

Each agent carried the recipe established earlier in this audit, including the
traps that had already cost attempts:

  a BARE named helper does not satisfy G115 -- the conversion is still flagged
    inside the wrapper, so reinterpretation needs an explicit branch that makes
    the range provable at each conversion
  byte(x >> n & 0xFF) passes; byte(x >> n) does not -- masking is not narrowing
  a range check applied directly to len(x) is not inferred; it has to be
    assigned to a local first, and `if len(x) < 0` separately trips staticcheck
    SA4024
  os.Root.ReadFile is clean for both constant and variable filenames, and
    refuses symlink/".." escape at the syscall layer
  prefer DELETING the conversion by carrying one type end-to-end -- that is what
    removed 14 of 26 findings in ee/kmip, and it is right more often than it
    looks

What remains is 15, and none of it is silenced:

  G204 x12  gosec accepts a subprocess only when EVERY argument is a literal. A
            literal command with a variable arg is still flagged, as are
            exec.LookPath and filepath.Join. These tests build into t.TempDir()
            and run a discovered openssl/python, so no literal form keeps them
            correct.
  G118 x1   ee/billing/install.go flags `go recorder.Run(ctx, ...)` for using
            context.Background(); it does not, ctx is threaded. gosec is wrong.
  G101 x1   a test fixture id matching on the FIELD name "credential".
  G304 x1   one site an agent reported as having no base directory to root at.

All four are on the operator list rather than waived on my own initiative.

.ee-lint-baseline drops 145 -> 15. The ratchet only ever moves down, so this is
now the ceiling.

Gates: go build ./... clean, go test ./ee/... rc=0, gen-cwe-docs.py current.
SLOW GATES DEFERRED to the end-of-batch run that follows immediately.

Finding: AH-57d4ba74

+++ ee/agentid/revoke/reconcile.go
@@ -100,7 +100,7 @@ func (c *Cascade) Reconcile(ctx context.Context, from uint64) (int, error) {
 				   (tenant_id, directive_id, subject_id, reason, watermark, terminal, seq)
 				 VALUES (current_setting('trstctl.tenant_id')::uuid, $1, $2, $3, $4, $5, $6)
 				 ON CONFLICT (tenant_id, directive_id) DO NOTHING`,
-				dir.DirectiveID, dir.SubjectID, dir.Reason, int64(dir.Watermark), dir.Terminal, int64(dir.Seq)); err != nil {
+				dir.DirectiveID, dir.SubjectID, dir.Reason, dir.Watermark, dir.Terminal, dir.Seq); err != nil {
 				return fmt.Errorf("revoke: reconcile directive row: %w", err)
 			}
 			for _, j := range jobRows {
